[Android] Update XWalkContentView to track ContentView
Move the methods for JellyBean to XWalkContentViewApi16 to imitate
JellyBeanContentView.
Move the mothods for AndroidM to XWalkContentViewApi23 to imitate
ContentView.ContentViewApi23.
All methods remaining in XWalkContentView is for XWalkView's usage.

XWalkContentView is a temporary solution. We are targeting to optimize
the hierarchy of XWalkView to make it more like WebView. This task can
be tracked via XWALK-6118.

BUG=XWALK-6014