[Android] Run Chromium's "sdkextras" hook as a Crosswalk hook.
The way the src/build/android/play_services/update.py script is called
from Crosswalk looks like this:
-> gclient sync
 -> hook invokes fetch_deps.py
  -> fetch_deps.py invokes gclient sync --gclientfile=.gclient-xwalk

Since we're stacking gclient on top of gclient, the process invocation
and I/O buffering code in depot_tools gets confused and the EULA prompt
is not shown at the right time. Users would normally just press Enter
when shown the EULA, which is then taken as "user has not accepted the
EULA".

I've had to invoke the update.py script directly from DEPS.xwalk as hook
instead. This does not cause any problems so far, we just have to keep
in mind that the script will be invoked after all other Chromium
hooks (in case the order becomes relevant in the future).

BUG=XWALK-5948